+++ src/app/embedded/exporter/exporter.component.ts
@@ -1,8 +1,8 @@
-import {Component, OnDestroy, OnInit, ViewChild} from '@angular/core';
+import {ChangeDetectorRef, Component, OnDestroy, OnInit, ViewChild} from '@angular/core';
 import {MatTableDataSource} from "@angular/material/table";
 import {MatPaginator} from "@angular/material/paginator";
 import {ExporterService, exportLog} from "../../teiler/exporter.service";
-import {from, Subscription} from "rxjs";
+import {from, map, Observable, Subscription} from "rxjs";
 import {Router} from "@angular/router";
 import {EmbeddedTeilerApps} from "../../teiler/teiler-app";
 import {QBResponse, Templates} from "../quality-report/quality-report.component";
@@ -10,6 +10,10 @@ import {environment} from "../../../environments/environment";
 import {SelectionModel} from "@angular/cdk/collections";
 import {TeilerAuthService} from "../../security/teiler-auth.service";
 import {createRouterLinkForBase} from "../../route/route-utils";
+import {FormBuilder, Validators} from "@angular/forms";
+import {BreakpointObserver} from "@angular/cdk/layout";
+import {StepperOrientation} from "@angular/cdk/stepper";
+import {ViewportScroller} from "@angular/common";
 
 export interface ExporterQueries {
   id: number;
@@ -21,6 +25,10 @@ export interface ExporterQueries {
   expirationDate: string;
   createdAt: string;
   archivedAt: string;
+  context: string;
+  defaultTemplateId: string;
+  defaultOutputFormat: string;
+
 }
 export interface ExportResponse {
   responseUrl: URL;
@@ -38,6 +46,10 @@ export enum ExportStatus {
   NOT_FOUND = "NOT_FOUND",
   EMPTY = "EMPTY",
   ERROR = "ERROR"
+}
+export interface Context {
+  key: string;
+  value: string;
 }
 @Component({
   selector: 'exporter',
@@ -57,16 +69,14 @@ export class ExporterComponent implements OnInit, OnDestroy {
   private subscriptionCreateQuery: Subscription | undefined;
   displayedColumns: string[] = ['#', 'timestamp', 'querytitle', 'querysource', 'format', 'executions'];
   dataSource = new MatTableDataSource<ExporterQueries>();
-  clickedRows = new Set<ExporterQueries>();
   buttonDisabled: boolean = true;
   editButtonDisabled: boolean = true;
   editModus: boolean = false;
-  exportLog: exportLog[] = [];
   templateIDs: Templates[] = [];
   outputFormats: DropdownFormat[] = [];
   queryFormats: DropdownFormat[] = [];
   selectedTemplate: string = environment.config.EXPORTER_DEFAULT_TEMPLATE_ID;
-  selectedOutputFormat: string = "JSON";
+  selectedOutputFormat: string = "EXCEL";
   selectedQueryFormat: string = "FHIR_SEARCH";
   exportUrl = "";
   fileName: string | undefined;
@@ -88,15 +98,41 @@ export class ExporterComponent implements OnInit, OnDestroy {
   queryList:ExporterQueries[] = [];
   executionLink: string = EmbeddedTeilerApps.EXECUTION;
   panelOpenState: boolean = false;
+  contextArray: Context[] = [{key: "", value: ""}];
+  showPlusButton: boolean = false;
 
-  constructor(private exporterService: ExporterService, private router: Router, public authService: TeilerAuthService) {
+
+  firstFormGroup = this._formBuilder.group({
+    queryTitle: [''],
+    queryDescription: ['']
+  });
+  secondFormGroup = this._formBuilder.group({
+    query: [''],
+    queryformat: [''],
+  });
+  thirdFormGroup = this._formBuilder.group({
+    template: [''],
+    outputformat: [''],
+  });
+  forthFormGroup = this._formBuilder.group({
+    expirationDate: [''],
+    contextKey: [''],
+    contextValue: ['']
+  });
+  stepperOrientation: Observable<StepperOrientation>;
+
+  constructor(private exporterService: ExporterService, private router: Router, public authService: TeilerAuthService, private _formBuilder: FormBuilder, breakpointObserver: BreakpointObserver, private viewport: ViewportScroller) {
     from(authService.loadUserProfile()).subscribe(keycloakProfile => this.contactID = keycloakProfile.email);
+    this.stepperOrientation = breakpointObserver
+      .observe('(min-width: 800px)')
+      .pipe(map(({matches}) => (matches ? 'horizontal' : 'vertical')));
   }
   // @ts-ignore
   @ViewChild('paginator') paginator: MatPaginator;
 
   ngAfterViewInit() {
     this.dataSource.paginator = this.paginator;
+    this.viewport.scrollToPosition([0,0]);
   }
   ngOnInit(): void {
     this.exportUrl = this.exporterService.getExporterURL() + "/";
@@ -118,7 +154,9 @@ export class ExporterComponent implements OnInit, OnDestroy {
     this.subscriptionCreateQuery?.unsubscribe();
     window.clearInterval(this.intervall);
   }
-
+  ngAfterContentChecked() {
+    //this.ref.detectChanges();
+  }
   getQueries(): void {
     this.subscriptionGetQueries?.unsubscribe();
     this.subscriptionGetQueries = this.exporterService.getReports().subscribe({
@@ -146,7 +184,10 @@ export class ExporterComponent implements OnInit, OnDestroy {
           contactId: query.contactId,
           expirationDate: this.transformDate(query.expirationDate),
           createdAt: this.transformDate(query.createdAt),
-          archivedAt: this.transformDate(query.archivedAt)
+          archivedAt: this.transformDate(query.archivedAt),
+          defaultTemplateId: query.defaultTemplateId,
+          defaultOutputFormat: query.defaultOutputFormat,
+          context: query.context
         });
       }
     })
@@ -163,9 +204,19 @@ export class ExporterComponent implements OnInit, OnDestroy {
     return new Date(date).getTime().toString();
   }
 
+  transformDateForQuery(date: Date | undefined): string {
+    if (date) {
+      const offset = date.getTimezoneOffset();
+      date = new Date(date.getTime() - (offset*60*1000));
+      return date.toISOString().split('T')[0];
+    } else {
+      return "";
+    }
+  }
+
   createAndExecuteQuery() {
     this.subscriptionGenerateExport?.unsubscribe();
-    this.subscriptionGenerateExport = this.exporterService.generateExport(this.query, this.queryLabel, this.queryDescription, this.selectedQueryFormat, this.selectedOutputFormat, this.selectedTemplate, this.importTemplate).subscribe({
+    this.subscriptionGenerateExport = this.exporterService.generateExport(this.query, this.queryLabel, this.queryDescription, this.selectedQueryFormat, this.selectedOutputFormat, this.selectedTemplate, this.getContext(), this.importTemplate).subscribe({
       next: (response: QBResponse) => {
         const url = new URL(response.responseUrl)
         const id = url.searchParams.get("query-execution-id");
@@ -202,15 +253,14 @@ export class ExporterComponent implements OnInit, OnDestroy {
           this.router.navigate([this.executionLink, this.loadedQueryID], {
             state: {
               newExecID: id,
-              query: this.query,
-              label: this.queryLabel,
-              description: this.queryDescription,
-              selectedQueryFormat: this.selectedQueryFormat,
-              selectedOutputFormat: this.selectedOutputFormat,
-              selectedTemplate: this.selectedTemplate
+              //query: this.query,
+              //label: this.queryLabel,
+              //description: this.queryDescription,
+              //selectedQueryFormat: this.selectedQueryFormat,
+              //selectedOutputFormat: this.selectedOutputFormat,
+              //selectedTemplate: this.selectedTemplate
             }
           })
-          //this.pollingStatusAndLogs(id, false);
         }
       },
       error: (error) => {
@@ -224,16 +274,17 @@ export class ExporterComponent implements OnInit, OnDestroy {
     this.subscriptionCreateQuery?.unsubscribe();
 
     this.buttonDisabled = true;
+    const expDate= this.transformDateForQuery(this.expirationDate);
 
-    if (this.executeOnSaving) {
-      this.createAndExecuteQuery();
-    } else {
       if (this.loadedQueryID !== "") {
-        this.subscriptionUpdateQuery = this.exporterService.updateQuery(this.loadedQueryID, this.queryLabel, this.queryDescription).subscribe({
+        this.subscriptionUpdateQuery = this.exporterService.updateQuery(this.loadedQueryID, this.query, this.queryLabel, this.queryDescription, this.selectedOutputFormat, this.selectedTemplate, this.getContext(), expDate, this.importTemplate).subscribe({
           next: (response: any) => {
             this.getQueries();
             this.editModus = false;
             this.buttonDisabled = false;
+            if (this.executeOnSaving) {
+              this.executeQuery();
+            }
           },
           error: (error) => {
             console.log(error);
@@ -243,11 +294,16 @@ export class ExporterComponent implements OnInit, OnDestroy {
           complete: () => {}
         });
       } else {
-        this.subscriptionCreateQuery = this.exporterService.createQuery(this.query, this.queryLabel, this.queryDescription, this.selectedQueryFormat, this.selectedOutputFormat, this.contactID, this.selectedTemplate, this.importTemplate).subscribe({
+        this.subscriptionCreateQuery = this.exporterService.createQuery(this.query, this.queryLabel, this.queryDescription, this.selectedQueryFormat, this.selectedOutputFormat, this.contactID, this.selectedTemplate, this.getContext(), expDate, this.importTemplate).subscribe({
           next: (response: QueryResponse) => {
             this.getQueries();
             this.editModus = false;
             this.buttonDisabled = false;
+            if (this.executeOnSaving) {
+              this.loadedQueryID = response.queryId;
+              this.executeQuery();
+            }
+
           },
           error: (error) => {
             console.log(error);
@@ -257,7 +313,7 @@ export class ExporterComponent implements OnInit, OnDestroy {
           complete: () => {}
         });
       }
-    }
+
   }
 
   getTemplateIDs(): void {
@@ -333,6 +389,7 @@ export class ExporterComponent implements OnInit, OnDestroy {
   }
   loadQuery(): void {
     const query: ExporterQueries[] = this.selection.selected
+
     if(query.length > 0) {
       this.query = query[0].query;
       this.queryDescription = query[0].description;
@@ -343,7 +400,21 @@ export class ExporterComponent implements OnInit, OnDestroy {
       this.executeOnSaving = true;
       this.buttonDisabled = false;
       this.loadedQueryID = query[0].id.toString();
+      query[0].defaultOutputFormat !== null && query[0].defaultOutputFormat !== undefined ? this.selectedOutputFormat = query[0].defaultOutputFormat : this.selectedOutputFormat = "EXCEL";
+      query[0].defaultTemplateId !== null && query[0].defaultTemplateId !== undefined ? this.selectedTemplate = query[0].defaultTemplateId : this.selectedTemplate = environment.config.EXPORTER_DEFAULT_TEMPLATE_ID;
       this.panelOpenState = true;
+      query[0].expirationDate !== "0" ? this.expirationDate = new Date(parseInt(query[0].expirationDate)) : this.expirationDate = undefined;
+      if (query[0].context !== null) {
+        this.contextArray = [];
+        atob(query[0].context).split(';').forEach((context) => {
+          const contextPair = context.split('=');
+          this.contextArray.push({key: contextPair[0], value: contextPair[1]} as Context);
+        })
+        this.showPlusButton = true;
+      } else {
+        this.contextArray = [{key: "", value: ""} as Context];
+      }
+
     }
     else {
       this.query = "";
@@ -354,6 +425,10 @@ export class ExporterComponent implements OnInit, OnDestroy {
       this.editButtonDisabled = true;
       this.buttonDisabled = true;
       this.loadedQueryID = "";
+      this.expirationDate = undefined;
+      this.selectedTemplate = environment.config.EXPORTER_DEFAULT_TEMPLATE_ID;
+      this.selectedOutputFormat = "EXCEL";
+      this.contextArray = [{key: "", value: ""} as Context];
     }
   }
 
@@ -364,11 +439,13 @@ export class ExporterComponent implements OnInit, OnDestroy {
     this.query = "";
     from(this.authService.loadUserProfile()).subscribe(keycloakProfile => this.contactID = keycloakProfile.email);
     this.selectedTemplate = environment.config.EXPORTER_DEFAULT_TEMPLATE_ID;
-    this.selectedOutputFormat = "JSON";
+    this.selectedOutputFormat = "EXCEL";
     this.selectedQueryFormat = "FHIR_SEARCH";
     this.expirationDate = undefined;
     this.loadedQueryID = "";
     this.panelOpenState = true;
+    this.contextArray = [{key: "", value: ""} as Context];
+    this.selection.clear();
     this.generateButtonStatus();
   }
 
@@ -386,4 +463,33 @@ export class ExporterComponent implements OnInit, OnDestroy {
   getRouterLink(id: string): string {
     return '/' + createRouterLinkForBase(this.executionLink + '/' + id);
   }
+  addContextInput(element: any, index2: number): void {
+    const scrollElement = element.target.parentNode.parentNode.parentNode;
+    this.contextArray.push({key: "", value: ""} as Context);
+    setTimeout(() => {
+      scrollElement.scrollTop = scrollElement.scrollHeight;
+    }, 50);
+    this.showPlusButton = false;
+  }
+  deleteContextInput(index: number): void {
+    this.contextArray.splice(index, 1);
+    this.checkContext(this.contextArray.length-1);
+  }
+  checkContext(index: number) {
+    this.showPlusButton = this.contextArray[index].key.length > 0 && this.contextArray[index].value.length > 0;
+  }
+  getContext(): string {
+    let context: string = "";
+    this.contextArray.forEach(contextPair => {
+      if (contextPair.key.length !== 0 && contextPair.value.length !== 0) {
+        if (context.length !== 0) {
+          context += ";"
+        }
+        context += contextPair.key + "=" + contextPair.value;
+      }
+    })
+    //return Buffer.from(context).toString("base64");
+    return btoa(context);
+  }
+
 }