+++ src/app/sidebar/sidebar.component.ts
@@ -3,6 +3,9 @@ import {TeilerService} from "../teiler/teiler.service";
 import {createMainRouterLink} from "../route/route-utils";
 import {ColorSchemeService} from "../color-scheme.service";
 import {RouteManagerService} from "../route/route-manager.service";
+import {TeilerApp} from "../teiler/teiler-app";
+import {environment} from "../../environments/environment";
+
 
 @Component({
   selector: 'app-sidebar',
@@ -15,8 +18,14 @@ export class SidebarComponent implements OnInit {
   public innerWidth: any;
   public menuVisibleMobile: boolean = false;
 
+  selectedBackground: string = 'background1';
+
   constructor(public teilerService: TeilerService, private colorSchemeService: ColorSchemeService, public routeManagerService: RouteManagerService) {
   }
+
+
+
+
   isGreyTheme(): boolean {
     return this.colorSchemeService.getColor() === 'rgb(211,211,211)';
   }
@@ -53,4 +62,12 @@ export class SidebarComponent implements OnInit {
     let mainRouterLink = createMainRouterLink();
     return window.location.pathname === '/' + mainRouterLink;
   }
+
+  isClickable(app:TeilerApp): boolean {
+    return !!((app.backendReachable && app.frontendReachable) ||
+      (app.backendReachable === undefined && app.frontendReachable) ||
+      (app.backendReachable === null && app.frontendReachable));
+  }
+
+  protected readonly environment = environment;
 }