+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.0.4] - 2025-03-04
+### Added
+- Assets API service
+- Default values for docker env for logo, background, color and font
+
 ## [1.0.2] - 2024-02-14
 ### Changed
 - OIDC instead of KEYCLOAK